% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{autohrf-datasets}

\alias{autohrf-datasets}
\alias{swm}
\alias{swm_autofit}
\alias{swm_autofit1}
\alias{swm_autofit2}
\alias{flanker}
\alias{flanker_autofit}

\title{Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.

The datasets were extracted from the internal Mind and Brain Lab's (MBLab, \url{http://www.mblab.si} repository. MBLab is a research lab at the Faculty of Arts, Department of Psychology, University of Ljubljana, Slovenia.}

\format{
\describe{
\item{\code{swm}}{
fMRI dataset for the spatial working memory study.

Source: Internal MBLab repository.

11520 obs. of 3 variables
\itemize{
\item \code{roi} region of interest.
\item \code{t} time stamp.
\item \code{y} BOLD value.
}
}

\item{\code{swm_autofit}}{
Stored results from a complete autohrf run for the spatial working memory study.

Source: Internal MBLab repository.
}

\item{\code{swm_autofit1}}{
Stored results from a complete autohrf run for a model with three event predictors (encoding, delay, and response) in the spatial working memory study.

Source: Internal MBLab repository.
}

\item{\code{swm_autofit2}}{
Stored results from a complete autohrf run for a model with four event predictors (encoding, early delay, late delay, and response) in the spatial working memory study.

Source: Internal MBLab repository.
}

\item{\code{flanker}}{
fMRI dataset for the flanker study.

Source: Internal MBLab repository.

192 obs. of 3 variables
\itemize{
\item \code{roi} region of interest.
\item \code{t} time stamp.
\item \code{y} BOLD value.
}
}

\item{\code{flanker_autofit}}{
Stored results from a complete autohrf run for the flanker study.

Source: Internal MBLab repository.
}

}
}

\description{
Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.

The datasets were extracted from the internal Mind and Brain Lab's (MBLab, \url{http://www.mblab.si} repository. MBLab is a research lab at the Faculty of Arts, Department of Psychology, University of Ljubljana, Slovenia.
}

\examples{
# load swm data
data_swm     <- swm
data_flanker <- flanker

# load the previously completed autofit
autofit_swm     <- swm_autofit
autofit_flanker <- flanker_autofit

}
