% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{autohrf}
\alias{autohrf}
\title{autohrf}
\usage{
autohrf(
  d,
  model_specs,
  population = 100,
  iter = 100,
  mutation_rate = 0.1,
  mutation_factor = 0.05,
  elitism = 0.1,
  tr = 2.5,
  f = 100,
  method = "middle",
  hrf = "boynton",
  t = 32,
  delta = 2.25,
  tau = 1.25,
  alpha = 2,
  p = c(6, 16, 1, 1, 6, 0, 32)
)
}
\arguments{
\item{d}{A dataframe with the signal data: roi, t and x. ROI is the name of the region, t timestamps and x values of the signal.}

\item{model_specs}{A list of model specifications to use for fitting. Each specification is represented as a data frame containing information about it (event, start_time, end_time, min_duration and max_duration).}

\item{population}{The size of the population for use in the genetic algorithm.}

\item{iter}{Number of iterations in the genetic algorithm.}

\item{mutation_rate}{The mutation rate in the genetic algorithm.}

\item{mutation_factor}{The mutation factor in the genetic algorithm.}

\item{elitism}{Whether to use elitism (promote a number of the best solutions) in the genetic algorithm.}

\item{tr}{MRI's repetition time.}

\item{f}{Downsampling frequency.}

\item{method}{Can be "middle" or "mean". Middle will return integer results, mean will return floats.}

\item{hrf}{Method to use for HRF generation.}

\item{t}{The t parameter for Boynton or SPM HRF generation.}

\item{delta}{The delta parameter of Boynton's HRF.}

\item{tau}{The tau parameter of Boynton's HRF.}

\item{alpha}{The alpha parameter of Boynton's HRF.}

\item{p}{The p parameter of SPM's HRF.}
}
\value{
A list containing model fits for each of the provided model specifications.
}
\description{
A function that automatically finds the parameters of model's that best match the underlying data.
}
\examples{
# prepare model specs
model3 <- data.frame(
  event        = c("encoding", "delay", "response"),
  start_time   = c(0,          2.65,     12.5     ),
  end_time     = c(3,          12.5,     16       ),
  min_duration = c(1,          5,        1        )
)

model4 <- data.frame(
  event        = c("fixation", "target", "delay", "response"),
  start_time   = c(0,          2.5,      2.65,    12.5),
  end_time     = c(2.5,        3,        12.5,    15.5),
  min_duration = c(1,          0.1,      5,       1)
)

model_specs <- list(model3, model4)

# run autohrf
df <- swm
autofit <- autohrf(df, model_specs, population=2, iter=2)

}
