% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{create_boynton_hrf}
\alias{create_boynton_hrf}
\title{create_boynton_hrf}
\usage{
create_boynton_hrf(tr, t = 32, p = c(2.25, 1.25, 2))
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{The t parameter for Boynton or SPM HRF generation.}

\item{p}{Parameters for the Boynton's HRF.}
}
\value{
Returns a Boynton HRF function.
}
\description{
A helper function for creating a Boynton HRF.
}
