% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{create_child}
\alias{create_child}
\title{create_child}
\usage{
create_child(
  start_time,
  end_time,
  n_events,
  mutation_rate,
  mutation_factor,
  current_model,
  p1,
  p2,
  allow_overlap
)
}
\arguments{
\item{start_time}{A list with model's event start times.}

\item{end_time}{A list with model's event end times.}

\item{n_events}{Number of events in the model.}

\item{mutation_rate}{The mutation rate in the genetic algorithm.}

\item{mutation_factor}{The mutation factor in the genetic algorithm.}

\item{current_model}{The constraints of the current model.}

\item{p1}{The first selected parent.}

\item{p2}{The second selected parent.}

\item{allow_overlap}{Whether to allow overlap between events.}
}
\value{
A child model created from two parents.
}
\description{
A helper function for creating a child from parents.
}
