% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{create_first_generation}
\alias{create_first_generation}
\title{create_first_generation}
\usage{
create_first_generation(current_model, n_events, population, allow_overlap)
}
\arguments{
\item{current_model}{The constraints of the current model.}

\item{n_events}{Number of events in the model.}

\item{population}{The size of the population in the genetic algorithm.}

\item{allow_overlap}{Whether to allow overlap between events.}
}
\value{
Returns the first generation of models.
}
\description{
A helper function for creating the first generation.
}
