% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{create_new_generation}
\alias{create_new_generation}
\title{create_new_generation}
\usage{
create_new_generation(
  elitism,
  population,
  start_time,
  end_time,
  fitness,
  n_events,
  mutation_factor,
  mutation_rate,
  current_model,
  allow_overlap
)
}
\arguments{
\item{elitism}{The degree of elitism (promote a percentage of the best
solutions) in the genetic algorithm.}

\item{population}{The size of the population in the genetic algorithm.}

\item{start_time}{A list with model's event start times.}

\item{end_time}{A list with model's event end times.}

\item{fitness}{A fitness score of all candidate models.}

\item{n_events}{Number of events in the model.}

\item{mutation_factor}{The mutation factor in the genetic algorithm.}

\item{mutation_rate}{The mutation rate in the genetic algorithm.}

\item{current_model}{The constraints of the current model.}

\item{allow_overlap}{Whether to allow overlap between events.}
}
\value{
A new generation of candidate models.
}
\description{
A helper function for creating a new generation of possible
solutions.
}
