% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{create_spm_hrf}
\alias{create_spm_hrf}
\title{create_boynton_hrf}
\usage{
create_spm_hrf(tr, t = 32, p = c(6, 16, 1, 1, 6, 0))
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{The t parameter for Boynton or SPM HRF generation.}

\item{p}{Parameters for the SPM HRF.}
}
\value{
Returns a SPM HRF function.
}
\description{
A helper function for creating a SPM HRF.
}
