% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{downsample}
\alias{downsample}
\title{downsample}
\usage{
downsample(y, f = 100)
}
\arguments{
\item{y}{The signal.}

\item{f}{Upsampling factor.}
}
\value{
Returns the downsampled signal.
}
\description{
A helper function for downsampling a given signal.
}
