% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{evaluate_model}
\alias{evaluate_model}
\title{evaluate_model}
\usage{
evaluate_model(
  d,
  model,
  roi_weights = NULL,
  normalize = TRUE,
  tr = 2.5,
  method = "middle",
  f = 100,
  hrf = "boynton",
  t = 32,
  delta = 2.25,
  tau = 1.25,
  alpha = 2,
  p = c(6, 16, 1, 1, 6, 0, 32)
)
}
\arguments{
\item{d}{A data frame with the signal data: roi, t and x.
ROI is the name of the region, t time stamps and x values of the signal.}

\item{model}{A data frame containing information about the model to use
and its events (event, start_time and duration).}

\item{roi_weights}{A data frame with ROI weights: roi, weight. ROI is the
name of the region, weight a number that defines the importance of that roi,
the default weight for a ROI is 1. If set to 2 for a particular ROI that ROI
will be twice as important.}

\item{normalize}{Whether to normalize the signal.}

\item{tr}{MRI's repetition time.}

\item{method}{Can be "middle" or "mean".
Middle will return integer results, mean will return floats.}

\item{f}{Downsampling frequency.}

\item{hrf}{Method to use for HRF generation, can be "boynton" or "spm".}

\item{t}{The t parameter for Boynton or SPM HRF generation.}

\item{delta}{The delta parameter of Boynton's HRF.}

\item{tau}{The tau parameter of Boynton's HRF.}

\item{alpha}{The alpha parameter of Boynton's HRF.}

\item{p}{The p parameter of SPM's HRF.}
}
\value{
Returns a list that contains the model, fits of events for
each ROI, convolved events and the TR.
}
\description{
A function for evaluating the model against the data.
}
\examples{
# create the model
m <- data.frame(event = c("encoding", "delay", "response"),
start_time = c(0, 2.5, 12.5), duration = c(2.5, 10, 5))

# convolve
ce <- convolve_events(m)

# evaluate
df <- swm
res <- evaluate_model(df, ce, m)

}
