% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_parents}
\alias{get_parents}
\title{get_parents}
\usage{
get_parents(fitness)
}
\arguments{
\item{fitness}{A fitness score of all candidate models.}
}
\value{
Parents for the child model.
}
\description{
A helper function for getting parents for the child model.
}
