% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_boynton}
\alias{hrf_boynton}
\title{hrf_boynton}
\usage{
hrf_boynton(tr = 2.5, t = 32, delta = 2.25, tau = 1.25, alpha = 2)
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{TODO}

\item{delta}{the delta parameter of Boynton's HRF.}

\item{tau}{the tau parameter of Boynton's HRF.}

\item{alpha}{the alpha parameter of Boynton's HRF.}
}
\value{
A vector representing a downsampled input signal.
}
\description{
Generate an HRF using the Boynton method.
}
\examples{
# create the HRF with a smaller tr
x <- hrf_boynton(tr=1)

}
