% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_spm}
\alias{hrf_spm}
\title{hrf_spm}
\usage{
hrf_spm(tr = 2.5, t = 16, p = c())
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{TODO}

\item{p}{TODO}
}
\value{
A vector representing a downsampled input signal.
}
\description{
Generate an HRF using the SPM method.
}
\examples{
# create the HRF with a smaller tr
x <- hrf_spm(tr=1)

}
