% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{plot_events}
\alias{plot_events}
\title{plot_events}
\usage{
plot_events(af, i = NULL)
}
\arguments{
\item{af}{The output from the autohrf function.}

\item{i}{Model index.}
}
\value{
Returns a plot of the events.
}
\description{
A helper function for plotting events of a fitted model.
}
