% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(model_evaluation)
}
\arguments{
\item{model_evaluation}{The output from the evaluate_model function.}
}
\description{
Plots a manually constructed model.
}
\examples{
# prepare model specs
model3 <- data.frame(event      = c("encoding", "delay", "response"),
                     start_time = c(0,           2.65,    12.5),
                     duration   = c(2.65,        9.85,    3))


}
