% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(d, r, model, normalize = TRUE, report = TRUE)
}
\arguments{
\item{d}{TODO.}

\item{r}{TODO.}

\item{model}{TODO.}

\item{normalize}{TODO.}

\item{report}{TODO.}
}
\value{
TODO.
}
\description{
TODO.
}
\examples{
# TODO
x <- "TODO"

}
