% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{autohrf-datasets}
\alias{autohrf-datasets}
\alias{swm}
\alias{swm_autofit}
\title{Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.
The datasets were extracted from the internal
Mind and Brain Lab's (MBLab, \url{http://www.mblab.si} repository.
MBLab is a research lab at the
Faculty of Arts, Department of Psychology,
University of Ljubljana, Slovenia.}
\format{
\describe{
\item{\code{swm}}{
fMRI dataset for a spatial working memory experiment.

Source: Internal MBLab repository.

504 obs. of 3 variables
\itemize{
\item \code{roi} region of interest.
\item \code{time} time stamp.
\item \code{y} BOLD value.
}
}

\item{\code{swm_autofit}}{
Stored results from a complete autohrf run.

Source: Internal MBLab repository.
}

}
}
\description{
Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.
The datasets were extracted from the internal
Mind and Brain Lab's (MBLab, \url{http://www.mblab.si} repository.
MBLab is a research lab at the
Faculty of Arts, Department of Psychology,
University of Ljubljana, Slovenia.
}
\examples{
# load swm data
data_swm <- swm

# load the previously completed autofit
autofit <- swm_autofit

}
