% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{autohrf-datasets}
\alias{autohrf-datasets}
\alias{swm}
\title{Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.
The datasets were extracted from the internal MBLab \url{http://www.mblab.si} repository.
MBLab is a research lab at the Faculty of Arts, Department of Psychology, University of Ljubljana, Slovenia.}
\format{
\describe{
\item{\code{swm}}{
fMRI dataset for a spatial working memory experiment.

Source: Internal MBLab repository.

504 obs. of 3 variables
\itemize{
\item \code{roi} region of interest.
\item \code{time} time stamp.
\item \code{x} BOLD value.
}
}
}
}
\description{
Datasets for autohrf examples
Example datasets for use in \pkg{autohrf} examples and vignettes.
The datasets were extracted from the internal MBLab \url{http://www.mblab.si} repository.
MBLab is a research lab at the Faculty of Arts, Department of Psychology, University of Ljubljana, Slovenia.
}
\examples{
# load swm data
data_swm <- swm

}
