% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{autohrf}
\alias{autohrf}
\title{autohrf}
\usage{
autohrf(
  d,
  models,
  population = 100,
  iter = 100,
  mutation_rate = 0.1,
  mutation_factor = 0.05,
  elitism = 0.1,
  tr = 2.5,
  f = 100,
  method = "middle",
  hrf = "boynton",
  t = 32,
  delta = 2.25,
  tau = 1.25,
  alpha = 2,
  p = c(6, 16, 1, 1, 6, 0, 32)
)
}
\arguments{
\item{d}{TODO.}

\item{models}{TODO.}

\item{population}{TODO.}

\item{iter}{TODO.}

\item{mutation_rate}{TODO.}

\item{mutation_factor}{TODO.}

\item{elitism}{TODO.}

\item{tr}{TODO.}

\item{f}{TODO.}

\item{method}{TODO.}

\item{hrf}{TODO.}

\item{t}{TODO.}

\item{delta}{TODO.}

\item{tau}{TODO.}

\item{alpha}{TODO.}

\item{p}{TODO.}
}
\value{
TODO.
}
\description{
TODO.
}
\examples{
# TODO
x <- "TODO"

}
