% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{convolve_events}
\alias{convolve_events}
\title{convolve_events}
\usage{
convolve_events(
  model,
  tr = 2.5,
  method = "middle",
  f = 100,
  hrf = "boynton",
  t = 32,
  delta = 2.25,
  tau = 1.25,
  alpha = 2,
  p = c(6, 16, 1, 1, 6, 0, 32)
)
}
\arguments{
\item{model}{TODO.}

\item{tr}{MRI's repetition time.}

\item{method}{TODO.}

\item{f}{TODO.}

\item{hrf}{TODO.}

\item{t}{TODO.}

\item{delta}{TODO.}

\item{tau}{TODO.}

\item{alpha}{TODO.}

\item{p}{TODO.}
}
\value{
TODO.
}
\description{
TODO.
}
\examples{
# TODO
x <- "TODO"

}
