% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{convolve_events}
\alias{convolve_events}
\title{convolve_events}
\usage{
convolve_events(
  model,
  tr = 2.5,
  method = "middle",
  f = 100,
  hrf = "boynton",
  t = 32,
  delta = 2.25,
  tau = 1.25,
  alpha = 2,
  p = c(6, 16, 1, 1, 6, 0, 32)
)
}
\arguments{
\item{model}{A data frame containing information about the model to use (event, start_time, duration).}

\item{tr}{MRI's repetition time.}

\item{method}{Can be "middle" or "mean". Middle will return integer results, mean will return floats.}

\item{f}{Downsampling frequency.}

\item{hrf}{Method to use for HRF generation, can be "boynton" or "spm".}

\item{t}{The t parameter for Boynton or SPM HRF generation.}

\item{delta}{The delta parameter of Boynton's HRF.}

\item{tau}{The tau parameter of Boynton's HRF.}

\item{alpha}{The alpha parameter of Boynton's HRF.}

\item{p}{The p parameter of SPM's HRF.}
}
\value{
A list containing convolved events, signal and time series.
}
\description{
Convolves events of a model with a generated HRF signal.
}
\examples{
# create the model
m <- data.frame(event = c("encoding", "delay", "response"),
start_time = c(0, 2.5, 12.5), duration = c(2.5, 10, 5))

# convolve
res <- convolve_events(m)

}
