% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{convolve_hrf}
\alias{convolve_hrf}
\title{convolve_hrf}
\usage{
convolve_hrf(x, hrf)
}
\arguments{
\item{x}{the input signal.}

\item{hrf}{HRF to convolve the signal with.}
}
\value{
A vector representing a downsampled input signal.
}
\description{
Convolve HRF with a signal.
}
\examples{
# create the HRF
hrf <- hrf_spm(tr=1)

# our signal
x <- sin(1:1000)

ts <- convolve_hrf(x, hrf)

}
