% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{convolve_hrf}
\alias{convolve_hrf}
\title{convolve_hrf}
\usage{
convolve_hrf(y, hrf_s)
}
\arguments{
\item{y}{The signal.}

\item{hrf_s}{The HRF.}
}
\value{
Returns the convolution between HRF and the signal.
}
\description{
A helper function for convolving HRF with a signal.
}
