% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{create_boynton_hrf}
\alias{create_boynton_hrf}
\title{create_boynton_hrf}
\usage{
create_boynton_hrf(tr = 2.5, t = 32, delta = 2.25, tau = 1.25, alpha = 2)
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{TODO}

\item{delta}{the delta parameter of Boynton's HRF.}

\item{tau}{the tau parameter of Boynton's HRF.}

\item{alpha}{the alpha parameter of Boynton's HRF.}
}
\value{
TODO.
}
\description{
Generate an HRF using the Boynton method.
}
\examples{
# create the HRF with a smaller tr
x <- create_boynton_hrf(tr=1)

}
