% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{create_spm_hrf}
\alias{create_spm_hrf}
\title{create_spm_hrf}
\usage{
create_spm_hrf(tr = 2.5, t = 32, p = c(6, 16, 1, 1, 6, 0, 32))
}
\arguments{
\item{tr}{MRI's repetition time.}

\item{t}{TODO}

\item{p}{TODO}
}
\value{
TODO.
}
\description{
Generate an HRF using the SPM method.
}
\examples{
# create the HRF with a smaller tr
x <- create_spm_hrf(tr=1)

}
