% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{downsample}
\alias{downsample}
\title{downsample}
\usage{
downsample(x, f = 100, method = "middle")
}
\arguments{
\item{x}{The indexes/time stamps to downsample.}

\item{f}{Downsampling frequency.}

\item{method}{Can be "middle" or "mean". Middle will return integer results, mean will return floats.}
}
\value{
A vector representing a downsampled input signal.
}
\description{
Downsample a time series by filtering its time stamps/indexes.
}
\examples{
# time stamps
x <- seq(1:1000)

# downsample
x_d <- downsample(x)

}
