% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(model, ce, tr = 2.5)
}
\arguments{
\item{model}{A data frame containing information about the model to use and
its events (event, start_time and duration).}

\item{ce}{The output of convolve_events by using the model and the data we
are testing against.}

\item{tr}{MRI's repetition time.}
}
\description{
Plots a manually constructed model.
}
\examples{
# prepare model specs
model3 <- data.frame(event      = c("encoding", "delay", "response"),
                     start_time = c(0,           2.65,    12.5     ),
                     duration   = c(2.65,        9.85,    3        ))


}
