% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(
  model_evaluation,
  by_roi = FALSE,
  ncol = NULL,
  nrow = NULL,
  scales = "free_y",
  rois = NULL
)
}
\arguments{
\item{model_evaluation}{The output from the evaluate_model function.}

\item{by_roi}{Whether to plot the fit for each ROI independently.}

\item{ncol}{Number of columns in the facet wrap.}

\item{nrow}{Number of rows in the facet wrap.}

\item{scales}{Whether to free certain axes of the facet wrap.}

\item{rois}{A subset of ROIs to visualize.}
}
\description{
Plots a manually constructed model.
}
\examples{
# prepare model specs
model3 <- data.frame(event      = c("encoding", "delay", "response"),
                     start_time = c(0,           2.65,    12.5),
                     duration   = c(2.65,        9.85,    3))


}
