% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print_best_models}
\alias{print_best_models}
\title{print_best_models}
\usage{
print_best_models(autofit)
}
\arguments{
\item{autofit}{Output of the autohrf function.}
}
\description{
Prints the best fitted model for each of the specs used in
autohrf.
}
\examples{
# prepare model specs
model3 <- data.frame(
  event        = c("encoding", "delay", "response"),
  start_time   = c(0,          2.65,     12.5),
  end_time     = c(3,          12.5,     16)
)

model4 <- data.frame(
  event        = c("fixation", "target", "delay", "response"),
  start_time   = c(0,          2.5,      2.65,    12.5),
  end_time     = c(2.5,        3,        12.5,    15.5)
)

model_specs <- list(model3, model4)

# run autohrf
df <- swm
autofit <- autohrf(df, model_specs, population=2, iter=2)

# print best models
print_best_models(autofit)

}
