% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(d, ce, model, roi_weights = NULL)
}
\arguments{
\item{d}{A dataframe with the signal data: roi, t and y. ROI is the name of
the region, t is the timestamp and y the value of the signal.}

\item{ce}{Result of the convolve_events function.}

\item{model}{A data frame containing information about the model to use
and its events (event, start_time and duration).}

\item{roi_weights}{A data frame with ROI weights: roi, weight. ROI is the
name of the region, weight a number that defines the importance of that roi,
the default weight for a ROI is 1. If set to 2 for a particular ROI that ROI
will be twice as important.}
}
\value{
Returns the model's evaluation.
}
\description{
A helper function for evaluating a model.
}
